refactor: remove MySpecies file
