+++ src/features/user/TreeMapper/MySpecies/MySpeciesForm.tsx
@@ -15,12 +15,22 @@ import { ErrorHandlingContext } from '../../../common/Layout/ErrorHandlingContex
 import { handleError, APIError } from '@planet-sdk/common';
 import InlineFormDisplayGroup from '../../../common/Layout/Forms/InlineFormDisplayGroup';
 import { Button, TextField } from '@mui/material';
+import { Species } from '../../../common/types/plantLocation';
+
+interface newSpecies {
+  aliases: string;
+  scientificSpecies: {
+    id: string;
+    name: string;
+    scientificName: string;
+  } | null;
+}
 
 export default function MySpeciesForm() {
   const { t } = useTranslation(['treemapper', 'me', 'common']);
   const { token, contextLoaded, logoutUser } = useUserProps();
   const { setErrors } = React.useContext(ErrorHandlingContext);
-  const [species, setSpecies] = React.useState<any[]>([]);
+  const [species, setSpecies] = React.useState<Species[]>([]);
   const [isUploadingData, setIsUploadingData] = React.useState(false);
 
   const defaultMySpeciesValue = {
@@ -39,7 +49,7 @@ export default function MySpeciesForm() {
 
   const fetchMySpecies = async () => {
     try {
-      const result = await getAuthenticatedRequest(
+      const result: Species[] = await getAuthenticatedRequest(
         '/treemapper/species',
         token,
         logoutUser
@@ -50,7 +60,7 @@ export default function MySpeciesForm() {
     }
   };
 
-  const deleteSpecies = async (id: number) => {
+  const deleteSpecies = async (id: string) => {
     try {
       await deleteAuthenticatedRequest(
         `/treemapper/species/${id}`,
@@ -63,14 +73,14 @@ export default function MySpeciesForm() {
     }
   };
 
-  const addSpecies = async (species: any) => {
+  const addSpecies = async (species: newSpecies) => {
     setIsUploadingData(true);
     const data = {
       aliases:
         species.aliases || species.aliases !== ''
           ? species.aliases
-          : species.scientificSpecies.name,
-      scientificSpecies: species.scientificSpecies.id,
+          : species.scientificSpecies?.name,
+      scientificSpecies: species.scientificSpecies?.id,
     };
     try {
       await postAuthenticatedRequest(
@@ -143,7 +153,7 @@ export default function MySpeciesForm() {
           </InlineFormDisplayGroup>
         </form>
         <div className={styles.mySpeciesContainer}>
-          {species.map((species: any) => {
+          {species.map((species: Species) => {
             return (
               <div key={species.id} className={styles.speciesContainer}>
                 <div className={styles.speciesName}>{species.aliases}</div>
